+++ packages/chrome-extension-remake/lib/drag.js
@@ -1,0 +1,50 @@
+import interact from "interactjs";
+
+let interactable = null;
+
+export function activate() {
+    if (interactable) {
+        return;
+    }
+    interactable = interact('*')
+        .draggable({
+            inertia: true,
+            modifiers: [
+                interact.modifiers.restrictRect({
+                    restriction: 'parent',
+                    endOnly: true
+                })
+            ],
+            autoScroll: true,
+            listeners: {
+                move: dragMoveListener,
+                end(event) {
+                    setMode("idle");
+                }
+            }
+        });
+}
+
+export function deactivate() {
+    if (!interactable) {
+        return;
+    }
+    interactable.unset();
+    interactable = null;
+}
+
+function dragMoveListener(event) {
+    setMode("dragging");
+    var target = event.target;
+    var x = (parseFloat(target.getAttribute('data-x')) || 0) + event.dx;
+    var y = (parseFloat(target.getAttribute('data-y')) || 0) + event.dy;
+    target.style.transform = 'translate(' + x + 'px, ' + y + 'px)';
+    target.setAttribute('data-x', x);
+    target.setAttribute('data-y', y);
+}
+
+export default {
+    activate,
+    deactivate,
+    dragMoveListener,
+};
