--- packages/chrome-extension-remake/lib/global.js
@@ -1,9 +1,0 @@
-export const state = {
-    mode: "",
-    target: null,
-    dragging: false,
-};
-
-export function setMode(mode) {
-    state.mode = mode;
-}
